+++ Sources/Swift/DataCollection/SentryDataCollection+DatabaseOptions.swift
@@ -15,5 +15,14 @@ extension SentryDataCollection {
         public init(queryParams: Bool = true) {
             self.queryParams = queryParams
         }
+
+        /// Creates database collection options from a dictionary.
+        @_spi(Private) public init(dictionary: [String: Any]) {
+            self.init()
+
+            if let queryParams = SentryDictionaryDecoder.bool(dictionary, "queryParams") {
+                self.queryParams = queryParams
+            }
+        }
     }
 }