+++ Sources/Swift/DataCollection/SentryDataCollection+GraphQLCollectionOptions.swift
@@ -24,5 +24,17 @@ extension SentryDataCollection {
             self.variables = variables
         }
 
+        /// Creates GraphQL collection options from a dictionary.
+        @_spi(Private) public init(dictionary: [String: Any]) {
+            self.init()
+
+            if let document = SentryDictionaryDecoder.bool(dictionary, "document") {
+                self.document = document
+            }
+            if let variables = SentryDictionaryDecoder.bool(dictionary, "variables") {
+                self.variables = variables
+            }
+        }
+
     }
 }