+++ Sources/Swift/DataCollection/SentryDataCollection.swift
@@ -1,3 +1,5 @@
+import Foundation
+
 /// Type used as namespace of all types related to data collection
 ///
 /// - SeeAlso: [Data Collection Spec](https://develop.sentry.dev/sdk/foundations/client/data-collection)