+++ Sources/Swift/SentryExperimentalOptions.swift
@@ -29,6 +29,11 @@ public final class SentryExperimentalOptions: NSObject {
     public var enableStandaloneAppStartTracing = false
 
     // swiftlint:disable:next missing_docs
-    @_spi(Private) public func validateOptions(_ options: [String: Any]?) {
+    @_spi(Private) public func populateFrom(dict options: [String: Any]?) {
+        guard let options else { return }
+
+        if let dataCollection = options["dataCollection"] as? [String: Any] {
+            self.dataCollection = SentryDataCollection.Options(dictionary: dataCollection)
+        }
     }
 }